backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
Pillow==10.0.1
Werkzeug==2.3.7
google-generativeai==0.3.2